Rename LeftPanel component to AppSidebar for clarity
